grid07-ai-assignment/requirements.txt
sentence-transformers
faiss-cpu
langchain
langgraph
python-dotenv
